refactored base game material/shader loading into a new class, called BaseGameTextureManager
