+++ XLGearModifier/AssetBundleHelper.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using TMPro;
 using UnityEngine;
 using XLGearModifier.CustomGear;
+using XLGearModifier.Texturing;
 using XLGearModifier.Unity;
 using Object = UnityEngine.Object;
 
@@ -22,7 +23,7 @@ namespace XLGearModifier
 
         public async Task LoadBundles()
         {
-            await GearManager.Instance.LoadGameMaterials();
+            await BaseGameTextureManager.Instance.LoadGameMaterials();
 
             // We're solely making a call here to ensure that the unity assembly is loaded up prior to loading assets.  else we'll get a bunch of errors about things missing.
             var test = GearModifierTab.CustomMeshes;
