+++ XLGearModifier/XLGearModifier.csproj
@@ -136,6 +136,7 @@
     <Compile Include="Patches\TextureMangingClassPatch.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
+    <Compile Include="Texturing\BaseGameTextureManager.cs" />
     <Compile Include="Texturing\SkateShopTextureManager.cs" />
     <Compile Include="UserInterfaceHelper.cs" />
   </ItemGroup>
